Introducing: Changelog.sh
It's quite simple, really. This creates a folder named DUChangelogs in
your build folder, and when ran, creates a subfolder with the date (mm-dd-yyyy).
Inside of said subfolder, it will output a file titled similarly to this:
DirtyUnicorns-Changelog-xx-xx-xxxx.log .    :)

Simple.
Organized.
Clean.

Original credits go to nuclearmistake , as I initially started with
one of his changelog scripts

ps2:
Got rid of the Carbon stuff (was for me, forgot to
remove it), removed all the whitespace I believe

ps3:
DERP

Change-Id: I28e2a73bf9c30b19129be9efceae1d9c96273214